refactor: Move all message generation logic to proper handler
